implement redeem wizard deposit status
